Convert scaling/instance int fields to uint64 for protobuf
These fields represent non-negative values (instance counts, CPU millicores,
memory MiB, in-flight requests) and need explicit sizing for future protobuf
definitions. Using `uint64` provides:

- Native `slog.Uint64()` support without conversion
- Direct protobuf compatibility with varint encoding
- Type-level enforcement that values cannot be negative

Removes now-dead validation code that checked for negative values, since
`uint64` naturally rejects them during JSON unmarshaling.

+++ internal/controller/pod.go
@@ -394,10 +394,10 @@ func (ctrl *Controller) instanceFromPod(pod *v1.Pod) (*function.Instance, error)
 	if podMetric, ok := ctrl.podMetrics.Load(pod.Namespace + "/" + pod.Name); ok {
 		for _, container := range podMetric.Containers {
 			if container.Usage.Cpu() != nil {
-				instance.CPUUsageMilli += int(container.Usage.Cpu().MilliValue())
+				instance.CPUUsageMilli += uint64(container.Usage.Cpu().MilliValue())
 			}
 			if container.Usage.Memory() != nil {
-				instance.MemoryUsageMiB += int(container.Usage.Memory().Value() / 1024 / 1024) // convert to MiB
+				instance.MemoryUsageMiB += uint64(container.Usage.Memory().Value() / 1024 / 1024) // convert to MiB
 			}
 		}
 	}
